Move api/crypto.h|c++ to api/crypto/crypto.h|c++
First step to reorganizing the crypto implementation in order to unify
the web crypto and node.js crypto implementation details into one set
of files with a clean implementation. The goal is to have a single set
of crypto implementation files that can be used easily by both APIs.

There will be some additional follow up commits to further clean things
up and move things around before implementing more of the stuff needed
to finish the Node.js crypto stuff.
