+++ src/workerd/api/api-rtti-test.c++
@@ -6,7 +6,7 @@
 #include <workerd/api/actor.h>
 #include <workerd/api/actor-state.h>
 #include <workerd/api/cache.h>
-#include <workerd/api/crypto.h>
+#include <workerd/api/crypto/crypto.h>
 #include <workerd/api/encoding.h>
 #include <workerd/api/global-scope.h>
 #include <workerd/api/html-rewriter.h>
