+++ src/workerd/api/crypto/crypto-impl-hmac.c++
@@ -3,7 +3,7 @@
 //     https://opensource.org/licenses/Apache-2.0
 
 #include "crypto-impl.h"
-#include <workerd/api/crypto.h>
+#include <workerd/api/crypto/crypto.h>
 #include <workerd/io/io-context.h>
 #include <openssl/hmac.h>
 #include <openssl/mem.h>
