+++ src/workerd/api/crypto/crypto.c++
@@ -4,12 +4,12 @@
 
 #include "crypto.h"
 #include "crypto-impl.h"
-#include "streams/standard.h"
+#include "../streams/standard.h"
+#include "../util.h"
 #include <array>
 #include <openssl/crypto.h>
 #include <openssl/err.h>
 #include <workerd/jsg/jsg.h>
-#include "util.h"
 #include <workerd/io/io-context.h>
 #include <workerd/util/uuid.h>
 #include <set>
